Removed duplicated code, method isValidTransition is encapsulated into a new object and job and query inherits from it
